POST method ready

+++ src/productManager.js
@@ -15,27 +15,28 @@ export default class ProductManager {
             thumbnail,
             code,
             stock,
-            id: this.autoincrementingId()
+            status: true,
+            id: await this.autoId(),
         }
         if(fs.existsSync(this.path)){
             fs.readFile(this.path, 'utf-8', (error, res)=>{
                 if(error) console.log(`Error: error in reading : ${error}`);
                 const products = JSON.parse(res);
-                /* if(products.some(existingProduct => existingProduct.code === product.code)) {
+                if(products.some(existingProduct => existingProduct.code === product.code)) {
                     console.log(`ERROR: this product code : ${product.code} already exists in products`);
                     console.log("--------------------------------");
-                } else if (title === "" || description === "" || price === "" || thumbnail === "" || code === "" || stock === "") {
+                } else if (title === "" || description === "" || price === "" || /* thumbnail === "" || thumbnail is commented because it is not obligatory in the challenge "primera entrega"*/code === "" || stock === "") {
                     console.log("Please complete all fields to push product")
                     console.log("--------------------------------");
-                } else { */
+                } else {
                     products.push(product);
                     fs.writeFile(this.path, JSON.stringify(products), (error)=> {
                         if(error) return console.log(`Error writing ${error}`);
                     })
-                /* } */
+                }
 
             })
-        }/* else{
+        }else{
             if(this.products.some(existingProduct => existingProduct.code === product.code)) {
             console.log(`ERROR: this product code : ${product.code} already exists in products`);
             console.log("--------------------------------");
@@ -52,40 +53,24 @@ export default class ProductManager {
                     })
                 })
             }
-        } */
+        }
     };
-    async autoincrementingId() {
-        if (fs.existsSync(this.path)) {
-          try {
-            const res = await fs.promises.readFile(this.path, 'utf-8');
-            const products = JSON.parse(res);
-            const lastId = products[products.length - 1];
-            this.id = lastId.id + 1;
-          } catch (error) {
-            console.log(`Error: error en la lectura del archivo: ${error}`);
-          }
-        } else {
-          if (this.products.length === 0) {
-            this.id = 1;
-          }
-        }
-        return this.id;
-      }
-    /* autoincrementingId() {
-        if(fs.existsSync(this.path)){
-            fs.readFile(this.path, 'utf-8', (error, res) => {
-                if (error) console.log(`Error : error in reading file : ${error}`);
-                const products = JSON.parse(res);
-                const lastId = products[products.length - 1];
-                products.id = lastId.id + 1;
-            })
-        }else {
-            if(this.products.length === 0) {
+    async autoId() {
+        try {
+            const products = await promises.readFile(this.path, 'utf-8');
+            const listProducts = JSON.parse(products);
+            
+            if (listProducts.length === 0) {
                 this.id = 1;
+            } else {
+                const lastId = listProducts[listProducts.length - 1].id;
+                this.id = lastId + 1;
             }
+            return this.id;
+        } catch (error) {
+            console.log("error in generate id", error);
         }
-        return this.id
-    } */
+    }
     async getProducts(){
         try {
             const products = await promises.readFile(this.path, 'utf-8');
